Add support for JNI
To avoid implementing the equivalent of RefMap I switched the XHandle
to a long. In a few places I skipped throwing exceptions. I did not
put too much time in teaching the gradle build system to build/pull
the .so/.h files. The top 'make java' is the one that does a correct
sequence of operations.